tried to fix the logo
not successful
