+++ app/src/main/res/layout/fragment_main.xml
@@ -7,14 +7,13 @@
 
 
     <ImageView
-        android:id="@+id/imageView2"
+        android:id="@+id/logo"
         android:layout_width="183dp"
         android:layout_height="196dp"
         android:layout_below="@+id/textView3"
         android:layout_centerHorizontal="true"
         android:layout_gravity="top|center"
         android:layout_marginTop="23dp"
-        android:visibility="visible"
         app:srcCompat="@mipmap/dragonfly" />
 
     <TextView
@@ -33,7 +32,7 @@
         android:id="@+id/new_trip"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_below="@+id/imageView2"
+        android:layout_below="@+id/logo"
         android:layout_centerHorizontal="true"
         android:layout_gravity="center"
         android:layout_marginTop="43dp"
@@ -49,15 +48,5 @@
         android:layout_marginTop="17dp"
         android:text="Saved Trips" />
 
-    <TextView
-
-        android:id="@+id/time"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top|center"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentStart="true"
-        android:layout_marginStart="14dp"
-        android:layout_marginBottom="34dp" />
 
 </RelativeLayout>
